chore: small code cleanup
